Fix build on gcc 4.9.3 + clean up warnings

+++ lib/threads.cpp
@@ -843,7 +843,7 @@ void communication_thread::execute()
          *is[0] >> ws >> filename;
 
          // all processors sent the screenshot command
-         for (int i = 1; i < is.size(); i++)
+         for (size_t i = 1; i < is.size(); i++)
          {
             *is[i] >> ws >> ident; // 'screenshot'
             *is[i] >> ws >> ident; // filename
@@ -861,7 +861,7 @@ void communication_thread::execute()
          *is[0] >> ws >> keys;
 
          // all processors sent the command
-         for (int i = 1; i < is.size(); i++)
+         for (size_t i = 1; i < is.size(); i++)
          {
             *is[i] >> ws >> ident; // 'keys'
             *is[i] >> ws >> ident; // keys
@@ -879,7 +879,7 @@ void communication_thread::execute()
          *is[0] >> w >> h;
 
          // all processors sent the command
-         for (int i = 1; i < is.size(); i++)
+         for (size_t i = 1; i < is.size(); i++)
          {
             *is[i] >> ws >> ident; // 'window_size'
             *is[i] >> t >> t;
@@ -897,7 +897,7 @@ void communication_thread::execute()
          *is[0] >> x >> y >> w >> h;
 
          // all processors sent the command
-         for (int i = 1; i < is.size(); i++)
+         for (size_t i = 1; i < is.size(); i++)
          {
             *is[i] >> ws >> ident; // 'window_geometry'
             *is[i] >> t >> t >> t >> t;
@@ -918,7 +918,7 @@ void communication_thread::execute()
          getline(*is[0], title, c);
 
          // all processors sent the command
-         for (int i = 1; i < is.size(); i++)
+         for (size_t i = 1; i < is.size(); i++)
          {
             *is[i] >> ws >> ident; // 'window_title'
             *is[i] >> ws >> c;
@@ -940,7 +940,7 @@ void communication_thread::execute()
          getline(*is[0], caption, c);
 
          // all processors sent the command
-         for (int i = 1; i < is.size(); i++)
+         for (size_t i = 1; i < is.size(); i++)
          {
             *is[i] >> ws >> ident; // 'plot_caption'
             *is[i] >> ws >> c;
@@ -966,7 +966,7 @@ void communication_thread::execute()
          getline(*is[0], label_z, c);
 
          // all processors sent the command
-         for (int i = 1; i < is.size(); i++)
+         for (size_t i = 1; i < is.size(); i++)
          {
             *is[i] >> ws >> ident; // 'axis_label'
             *is[i] >> ws >> c;
@@ -987,7 +987,7 @@ void communication_thread::execute()
       else if (ident == "pause")
       {
          // all processors sent the command
-         for (int i = 1; i < is.size(); i++)
+         for (size_t i = 1; i < is.size(); i++)
          {
             *is[i] >> ws >> ident; // 'pause'
          }
@@ -1004,7 +1004,7 @@ void communication_thread::execute()
          *is[0] >> theta >> phi;
 
          // all processors sent the command
-         for (int i = 1; i < is.size(); i++)
+         for (size_t i = 1; i < is.size(); i++)
          {
             *is[i] >> ws >> ident; // 'view'
             *is[i] >> a >> a;
@@ -1022,7 +1022,7 @@ void communication_thread::execute()
          *is[0] >> factor;
 
          // all processors sent the command
-         for (int i = 1; i < is.size(); i++)
+         for (size_t i = 1; i < is.size(); i++)
          {
             *is[i] >> ws >> ident; // 'zoom'
             *is[i] >> a;
@@ -1040,7 +1040,7 @@ void communication_thread::execute()
          *is[0] >> tot >> bdr;
 
          // all processors sent the command
-         for (int i = 1; i < is.size(); i++)
+         for (size_t i = 1; i < is.size(); i++)
          {
             *is[i] >> ws >> ident; // 'subdivisions'
             *is[i] >> a >> a;
@@ -1058,7 +1058,7 @@ void communication_thread::execute()
          *is[0] >> minv >> maxv;
 
          // all processors sent the command
-         for (int i = 1; i < is.size(); i++)
+         for (size_t i = 1; i < is.size(); i++)
          {
             *is[i] >> ws >> ident; // 'valuerange'
             *is[i] >> a >> a;
@@ -1076,7 +1076,7 @@ void communication_thread::execute()
          *is[0] >> ws >> shd;
 
          // all processors sent the command
-         for (int i = 1; i < is.size(); i++)
+         for (size_t i = 1; i < is.size(); i++)
          {
             *is[i] >> ws >> ident; // 'shading'
             *is[i] >> ws >> ident;
@@ -1094,7 +1094,7 @@ void communication_thread::execute()
          *is[0] >> x >> y;
 
          // all processors sent the command
-         for (int i = 1; i < is.size(); i++)
+         for (size_t i = 1; i < is.size(); i++)
          {
             *is[i] >> ws >> ident; // 'viewcenter'
             *is[i] >> a >> a;
@@ -1112,7 +1112,7 @@ void communication_thread::execute()
          *is[0] >> ws >> mode;
 
          // all processors sent the command
-         for (int i = 1; i < is.size(); i++)
+         for (size_t i = 1; i < is.size(); i++)
          {
             *is[i] >> ws >> ident; // 'autoscale'
             *is[i] >> ws >> ident;
@@ -1130,7 +1130,7 @@ void communication_thread::execute()
          *is[0] >> pal;
 
          // all processors sent the command
-         for (int i = 1; i < is.size(); i++)
+         for (size_t i = 1; i < is.size(); i++)
          {
             *is[i] >> ws >> ident; // 'palette'
             *is[i] >> a;
@@ -1148,7 +1148,7 @@ void communication_thread::execute()
          *is[0] >> n;
 
          // all processors sent the command
-         for (int i = 1; i < is.size(); i++)
+         for (size_t i = 1; i < is.size(); i++)
          {
             *is[i] >> ws >> ident; // 'palette_repeat'
             *is[i] >> a;
@@ -1169,7 +1169,7 @@ void communication_thread::execute()
          }
 
          // all processors sent the command
-         for (int i = 1; i < is.size(); i++)
+         for (size_t i = 1; i < is.size(); i++)
          {
             *is[i] >> ws >> ident; // 'camera'
             for (int j = 0; j < 9; j++)
@@ -1190,7 +1190,7 @@ void communication_thread::execute()
          *is[0] >> ws >> mode;
 
          // all processors sent the command
-         for (int i = 1; i < is.size(); i++)
+         for (size_t i = 1; i < is.size(); i++)
          {
             *is[i] >> ws >> ident; // 'autopause'
             *is[i] >> ws >> ident;
@@ -1210,7 +1210,7 @@ void communication_thread::execute()
    cout << "Stream: end of input." << endl;
 
 comm_terminate:
-   for (int i = 0; i < is.size(); i++)
+   for (size_t i = 0; i < is.size(); i++)
    {
       socketstream *isock = dynamic_cast<socketstream *>(is[i].get());
       if (isock)
